LayoutTests/fast/repaint/outline-transition-expected.txt

Test that using a CSS Transition to animate the 'outline' property causes repaints.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals?.repaintRectsAsText().indexOf('(rect 0 0 140 140)') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
